// frontend/src/common/models/measurement/Measurement.ts
import { BaseModel, transformerDateTime } from 'src/common/models/models';
import { Transform } from 'class-transformer';
import dayjs from 'dayjs';

export class Measurement extends BaseModel {
  sensorName?: string;
  readOrder?: number;

  accelMssX?: number;
  accelMssY?: number;
  accelMssZ?: number;

  accelLinX?: number;
  accelLinY?: number;
  accelLinZ?: number;

  gyroRadsX?: number;
  gyroRadsY?: number;
  gyroRadsZ?: number;

  magBiasX?: number;
  magBiasY?: number;
  magBiasZ?: number;

  roll?: number;
  pitch?: number;
  yaw?: number;

  eulerX?: number;
  eulerY?: number;
  eulerZ?: number;

  quaternionX?: number;
  quaternionY?: number;
  quaternionZ?: number;
  quaternionW?: number;

  @Transform(transformerDateTime)
  capturedAt?: dayjs.Dayjs;
}